+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.20 - Mar 23 2019
 - Improved email regex.
 - Display the correct error message when confirmation token is invalid
+- Add option `--skip-dependencies` to the commands `deploy-to` and `deploy` to allow skipping of `bundle install` and `shards install`.
 
 ## 0.5.19 - Mar 20 2019
 - Fix bug preventing `faastruby deploy` from working when secrets are nil